% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{otter_api}
\alias{otter_api}
\title{Otter API starting point}
\usage{
otter_api(
  api_token,
  base_api_url = "https://otter.ccm.sickkids.ca/api",
  base_app_url = "https://otter.ccm.sickkids.ca/app"
)
}
\arguments{
\item{api_token}{Your API token from https://otter.ccm.sickkids.ca/app/api_docs.}

\item{base_api_url}{API link.}

\item{base_app_url}{App link.}
}
\value{
An object to be passed to other functions.
}
\description{
This function initializes the Otter API.
}
