% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{otter_run_sample_path}
\alias{otter_run_sample_path}
\title{Run sample through the API}
\usage{
otter_run_sample_path(
  otter_obj,
  file_path,
  model_name,
  atlas_version = NULL,
  sample_name = NULL,
  share_with = NULL,
  wait_for_result = TRUE,
  timeout = 300
)
}
\arguments{
\item{otter_obj}{Your API object.}

\item{file_path}{path to TPM counts file.}

\item{model_name}{otter or hierarchical.}

\item{atlas_version}{not used for now, but kept here for compatibility with future versions of the atlas.}

\item{sample_name}{Name of the sample.}

\item{share_with}{List of emails to share the sample with.}

\item{wait_for_result}{Whether to submit the job and move on or wait for the results.}

\item{timeout}{When waiting, time to timeout.}
}
\value{
Dataframe of prediction or task id and inference id.
}
\description{
Runs a sample from a path to their TPMs.
}
